[F#] Use float instead of double (#1531)

+++ languages/fsharp/exercises/concept/numbers/Numbers.fs
@@ -1,5 +1,7 @@
 module Numbers
 
-let productionRatePerHour (speed: int): double = failwith "Please implement the 'productionRatePerHour' function"
+let productionRatePerHour (speed: int): float =
+    failwith "Please implement the 'productionRatePerHour' function"
 
-let workingItemsPerMinute (speed: int): int = failwith "Please implement the 'workingItemsPerMinute' function"
+let workingItemsPerMinute (speed: int): int =
+    failwith "Please implement the 'workingItemsPerMinute' function"
